requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.4
idna==2.10
isort==5.6.4
lazy-object-proxy==1.4.3
marshmallow==3.9.1
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
wrapt==1.12.1
